Backend/requirements.txt
Flask==2.2.5
SQLAlchemy==1.4.49
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
gunicorn==20.1.0
APScheduler==3.10.4

# Database migration (optional)
alembic==1.13.1

# Job scraping
jobspy

# Cross-origin helper for Flask (used in development)
Flask-Cors==4.0.0

# Data validation
pydantic==2.6.4
